Initial commit
** Notes **
Hyper-V's MMC snap-in defaults to creating automatic snapshots.

The current version of the scripts contained here in aren't really all
that prepared to deal correctly with snapshots/checkpoints, so New-Toy disables
automatic checkpoints.
And although Get-OrphanedVHDs and Remove-Toy contain code for removing
snapshots, the code has shown to be flaky because of timing issues since
snapshots seem to be merged asynchronously resulting in the following
instructions being executed before the merge being done.