Added support for SecureBoot and removed PXE
Added a new optional parameter for informing a Secure Boot template:
$SecureBootTemplate

For guest OSes that support Secure Boot the template can be specifed.
For instance: MicrosoftUEFICertificateAuthority

For guest OSes that don't support, the parameter should not be provided

The network adapter was removed from Boot Order since I'm not yet using
PXE. This should speed up the initial boot for installing the OS as the
the firmware will jump from the harddrive directly to the DVD drive
without having to timeout on the network adapter.