Fixed snapshot bug
Previous revision got the file name from the VM.
If the VM had a snapshot, the name would point to a snapshot file.
Later when the VM was removed, PowerShell automatically removed any
existing snapshots merging them into the root VHD.
Then the script tried to delete the file (snapshot) which the VM was
pointing to which at this point didn't exist any more.
As a result, the original (root) VHD was left behind untouched.