Added proper support for checkpoints/snapshots
The previous version of the script removed the snapshots of all virtual
machines in the process of creating a list of files being used by those
virtual machines, as I hadn't yet figured out how to link the VMs to any
but the last snapshot.
It was ugly but worked in a world where the only existing snapshots were
there by "accident". It just happens that the default Hyper-V seetings
turn on automatic checkpoints.
The new version creates a list of VHDs linked to the VMs themselves plus
VHDs linked to any snapshots that may exist.